Update to gcc 2.95.1, official release.

+++ gnu/egcs/gcc/fixinc/inclhack.sh
@@ -2331,6 +2331,7 @@ extern unsigned int\
         -e 's/char[ 	]*\*[ 	]*calloc/void \*	calloc/g' \
         -e 's/char[ 	]*\*[ 	]*malloc/void \*	malloc/g' \
         -e 's/char[ 	]*\*[ 	]*realloc/void \*	realloc/g' \
+        -e 's/char[ 	]*\*[ 	]*bsearch/void \*	bsearch/g' \
         -e 's/int[ 	][ 	]*exit/void	exit/g' \
         -e '/typedef[ 	a-zA-Z_]*[ 	]size_t[ 	]*;/i\
 #ifndef _GCC_SIZE_T\
@@ -3093,8 +3094,10 @@ done
 #
 # # # # # # # # # # # # # # # # # # # # #
 
-cd $ORIGDIR
-rm -f include/assert.h
-cp ${srcdir}/assert.h include/assert.h || exit 1
-chmod a+r include/assert.h
-
+if [ x${INSTALL_ASSERT_H} != x ] && [ -f ${srcdir}/assert.h ]
+then
+  cd $ORIGDIR
+  rm -f include/assert.h
+  cp ${srcdir}/assert.h include/assert.h || exit 1
+  chmod a+r include/assert.h
+fi
